Use cocoapods instead of library
