+++ .gitignore
@@ -16,6 +16,7 @@ DerivedData
 *.perspectivev3
 !default.perspectivev3
 xcuserdata
+.DS_Store
 
 ## Other
 *.xccheckout
