+++ .travis.yml
@@ -1,7 +1,14 @@
+# references:
+# * http://www.objc.io/issue-6/travis-ci.html
+# * https://github.com/supermarin/xcpretty#usage
+
+osx_image: xcode7.3
 language: objective-c
-osx_image: xcode7.2
+# cache: cocoapods
+podfile: Example/Podfile
+# before_install:
+# - gem install cocoapods # Since Travis is not always on latest version
+# - pod install --project-directory=Example
 script:
-    - xctool -sdk iphonesimulator -project Polymorph.xcodeproj -scheme "PolymorphTests iOS" test GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES
-    - xctool -sdk macosx -project Polymorph.xcodeproj -scheme "PolymorphTests OSX" test GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES
-after_success:
-    - bash <(curl -s https://codecov.io/bash)
+- set -o pipefail && xcodebuild test -workspace Example/Polymorph.xcworkspace -scheme Polymorph-Example -sdk iphonesimulator9.3 ONLY_ACTIVE_ARCH=NO | xcpretty
+- pod lib lint
