+++ Example/Podfile
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+platform :ios, '8.0'
+use_frameworks!
+
+target 'Polymorph_Example' do
+  pod 'Polymorph', :path => '../'
+
+  target 'Polymorph_Tests' do
+    inherit! :search_paths
+  end
+end
+
