+++ Example/Polymorph/Bridge-Swift.h
