+++ Example/Polymorph/Images.xcassets/AppIcon.appiconset/120.png
