+++ Example/Polymorph/Images.xcassets/AppIcon.appiconset/87.png
