+++ Example/Polymorph/Images.xcassets/Contents.json
